requirements.txt
langchain>=0.2.14
langchain-google-genai>=1.0.7
langchain-openai>=0.1.22
langchain-mistralai>=0.1.12
langchain_community>=0.2.9
langchain-aws>=0.1.3
html2text>=2024.2.26
faiss-cpu>=1.8.0
beautifulsoup4>=4.12.3
pandas>=2.2.2
python-dotenv>=1.0.1
tiktoken>=0.7
tqdm>=4.66.4
minify-html>=0.15.0
free-proxy>=1.1.1
playwright>=1.43.0
undetected-playwright>=0.3.0
semchunk>=1.0.1
langchain-ollama>=0.1.3
simpleeval>=0.9.13
googlesearch-python>=1.2.5 
async_timeout>=4.0.3
streamlit
ipython
ipykernel
jupyter
jupyterlab
pandas